Support plantuml as executable in PATH
The plantuml JAR is often packaged alongside a binary that calls the
JAR - both in OS X Homebrew and in Debian.  In particular, in OS X the
plantuml binary sets the environment so that it "just works" - hence
on OS X the wrapper is the recommended way of calling plantuml.

Sample wrapping binaries in packaged plantuml:
* [OS X homebrew](https://github.com/Homebrew/homebrew-core/blob/6c8c13b890dd3271d4fa05a08f5fc67ebf6055df/Formula/plantuml.rb#L14-L17)
* [debian](https://packages.debian.org/search?searchon=contents&keywords=plantuml&mode=exactfilename&suite=testing&arch=any)

Also convert shell-command to call-process.

References:
* https://www.gnu.org/software/emacs/manual/html_mono/elisp.html
* http://stackoverflow.com/questions/23299314/finding-the-exit-code-of-a-shell-command-in-elisp/23299809#comment35678383_23299809
* http://emacs.stackexchange.com/questions/7712/bind-multiple-values-directly-from-list-without-binding-the-list-itself