“fix” for #62: Unable to change any colors
This is not really a fix as this problem is not a problem of
plantuml-mode!

The generated svg (generated by plantuml) is displayed by several svg
viewers with a transparent background. One of such viewers is
ImageMagick's display command (which is used by emacs if it is
present). On the other hand, most web-browsers show it on a coloured
background. The problem can be solved by plantuml such, that they
insert a <rect width="100%" height="100%" fill="#A9A9A9"/> as first
child of the svg-element.

However, plantuml-mode can pass png as the default image type to
plantuml. In this case users of plantuml-mode will see the correct result.