Allow reinitialization of font lock regexps
In `plantuml-mode`, initialization of the font locking regular expressions is
done by asking the current version of plantuml for its supported keywords,
builtins, etc.  Every so often, this might fail for any reason, leaving the
current Emacs session with invalid font lock information for plantuml.  Redoing
the initialization process is tedious, as the corresponding function
`plantuml-init-once` does not allow for that to be done easily (one would have
to undefine a lot of global variables to do so).

This commit introduces a force parameter for `plantuml-init-once` to conduct
reinitialization of all font lock information.  It also moves regular
expressions meant for indentation out of `plantuml-init-once`, as those are
initialized by literal values that do not depend on any runtime information.