Move point at most to beginning of line when indenting
When indenting a line, point is moved synchronously to keep the current position
relative to the surrounding text.  This works well when the indentation level is
increased (point is moved to the right then), but when the indentation level is
decreased (point is moved to the left), it might happen that point is moved
beyond the beginning of the current line, positioning it somewhere on the
previous line or even before that.  This commit fixes this undesired behavior by
never moving point before the beginning of the current line.