feat: Perform the computation in a separate buffer before showing it
The previous implementation first killed the preview buffer and
then run plantuml. If plantuml takes a few seconds, it is annoying
to have the buffer vanishing and appearing again a few seconds
later.

It seems smoother to perform the computation while still showing
the obsolete buffer and replace the buffer when the result is
ready.